refactor: dedup colonizethis_map render layer and test harnesses (#3846)
Split tile_map_visualization into legend/overlay modules, consolidate view-builder
and determinism tests around shared fixtures, extend repo lint gates for render
size and topology dedup, and rename map_gen_file_size to map_lib_file_size.

Refs #3846
