--- packages/colonizethis_map/test/init_game_map_view_builder_warp_markers_test.dart
@@ -1,117 +1,0 @@
-import 'package:colonizethis_test/test.dart';
-import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_map/colonizethis_map.dart';
-
-import 'support/init_game_map_view_fixtures.dart';
-
-void main() {
-  group('buildInitGameMapViewData warp zone markers', () {
-    test('includes warp zone markers from warpLinks (bidirectional)', () {
-      final owMap = mapTileGrid([
-        ['s1', 's2', 's3'],
-      ]);
-      final nwMap = mapTileGrid([
-        ['s1', 's2', 's3'],
-      ]);
-      final owTopology = regionTopology(
-        regionId: 'oldWorld',
-        seaZoneIds: const ['s1', 's2', 's3'],
-      );
-      final nwTopology = regionTopology(
-        regionId: 'newWorld',
-        seaZoneIds: const ['s1', 's2', 's3'],
-      );
-      final game = minimalGame(id: 'warp');
-
-      // Warp links as generated by the generator: one-directional from OW to NW.
-      // The builder handles both directions (source and destination regions).
-      final warpLinks = [
-        WarpLink(
-          regionId: 'oldWorld',
-          seaZoneId: 's1',
-          otherRegionId: 'newWorld',
-          otherSeaZoneId: 's3',
-        ),
-        WarpLink(
-          regionId: 'oldWorld',
-          seaZoneId: 's2',
-          otherRegionId: 'newWorld',
-          otherSeaZoneId: 's2',
-        ),
-      ];
-
-      final viewData = buildInitGameMapViewData(
-        game: game,
-        tileMapByRegion: {'oldWorld': owMap, 'newWorld': nwMap},
-        topologyByRegion: {'oldWorld': owTopology, 'newWorld': nwTopology},
-        cellSize: 8,
-        warpLinks: warpLinks,
-      );
-
-      // Old World should have 2 warp markers (s1 and s2).
-      expect(viewData.oldWorld.warpMarkers, hasLength(2));
-      final s1Marker = viewData.oldWorld.warpMarkers.singleWhere(
-        (m) => m.seaZoneId == 's1',
-      );
-      expect(s1Marker.x, 0); // s1 is at tile index 0
-      expect(s1Marker.y, 0);
-      expect(s1Marker.otherRegionId, 'newWorld');
-      expect(s1Marker.otherSeaZoneId, 's3');
-
-      final s2Marker = viewData.oldWorld.warpMarkers.singleWhere(
-        (m) => m.seaZoneId == 's2',
-      );
-      expect(s2Marker.x, 1); // s2 is at tile index 1
-      expect(s2Marker.y, 0);
-      expect(s2Marker.otherRegionId, 'newWorld');
-      expect(s2Marker.otherSeaZoneId, 's2');
-
-      // New World should have 2 warp markers (s3 and s2) via reverse lookup.
-      expect(viewData.newWorld.warpMarkers, hasLength(2));
-      final nwS3Marker = viewData.newWorld.warpMarkers.singleWhere(
-        (m) => m.seaZoneId == 's3',
-      );
-      expect(nwS3Marker.x, 2); // s3 is at tile index 2
-      expect(nwS3Marker.y, 0);
-      expect(nwS3Marker.otherRegionId, 'oldWorld');
-      expect(nwS3Marker.otherSeaZoneId, 's1');
-
-      final nwS2Marker = viewData.newWorld.warpMarkers.singleWhere(
-        (m) => m.seaZoneId == 's2',
-      );
-      expect(nwS2Marker.x, 1); // s2 is at tile index 1
-      expect(nwS2Marker.y, 0);
-      expect(nwS2Marker.otherRegionId, 'oldWorld');
-      expect(nwS2Marker.otherSeaZoneId, 's2');
-    });
-
-    test('empty warpMarkers when warpLinks is null', () {
-      final owMap = mapTileGrid([
-        ['s1'],
-      ]);
-      final nwMap = mapTileGrid([
-        ['s1'],
-      ]);
-      final owTopology = regionTopology(
-        regionId: 'oldWorld',
-        seaZoneIds: const ['s1'],
-      );
-      final nwTopology = regionTopology(
-        regionId: 'newWorld',
-        seaZoneIds: const ['s1'],
-      );
-      final game = minimalGame(id: 'no-warp');
-
-      final viewData = buildInitGameMapViewData(
-        game: game,
-        tileMapByRegion: {'oldWorld': owMap, 'newWorld': nwMap},
-        topologyByRegion: {'oldWorld': owTopology, 'newWorld': nwTopology},
-        cellSize: 8,
-        warpLinks: null,
-      );
-
-      expect(viewData.oldWorld.warpMarkers, isEmpty);
-      expect(viewData.newWorld.warpMarkers, isEmpty);
-    });
-  });
-}
