+++ packages/colonizethis_map/test/support/init_game_map_view_fixtures.dart
@@ -1,4 +1,5 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 /// Shared scaffolding for the `buildInitGameMapViewData` view-builder test suite
@@ -105,5 +106,125 @@ Game minimalGame({
     players: players,
     minorNations: minorNations,
     tribes: tribes,
+  );
+}
+
+/// Dual-region topology + tile maps used by visualization render tests.
+MapTopology oldWorldTwoProvinceSeaVisualizationTopology() {
+  return regionTopology(
+    regionId: 'oldWorld',
+    provinceIds: const ['p1', 'p2'],
+    seaZoneIds: const ['s1'],
+    edges: const [
+      TopologyEdge(id1: 'p1', id2: 'p2'),
+      TopologyEdge(id1: 'p1', id2: 's1'),
+    ],
+  );
+}
+
+/// Two adjacent sea zones in [regionId] (sea–sea border render tests).
+MapTopology twoAdjacentSeaZonesTopology(String regionId) {
+  return regionTopology(
+    regionId: regionId,
+    seaZoneIds: const ['s1', 's2'],
+    edges: const [TopologyEdge(id1: 's1', id2: 's2')],
+  );
+}
+
+/// Standard 4×3 visualization grid with p1/p2 land and s1 sea.
+TileMapResult visualizationSmallTileMap() {
+  return mapTileGrid([
+    ['p1', 'p1', 'p2', 'p2'],
+    ['p1', 's1', 's1', 'p2'],
+    ['p1', 'p1', 'p2', 'p2'],
+  ]);
+}
+
+/// Dual-region view-builder / visualizer scenario wiring.
+class DualRegionViewScenario {
+  const DualRegionViewScenario({
+    required this.game,
+    required this.tileMapByRegion,
+    required this.topologyByRegion,
+  });
+
+  final Game game;
+  final Map<String, TileMapResult> tileMapByRegion;
+  final Map<String, MapTopology> topologyByRegion;
+
+  InitGameMapViewData buildViewData({
+    int cellSize = 8,
+    int? seed,
+    String? configSummary,
+    Map<String, (int r, int g, int b)>? greatPowerColorOverride,
+    Map<String, TileVisibility>? visibilityByTile,
+    List<WarpLink>? warpLinks,
+    Map<String, int>? resourceExtractionUnitsByTile,
+    Map<String, int>? resourceExtractionEffectiveUnitsByTile,
+    Map<String, int>? resourceExtractionBlockedUnitsByTile,
+    Set<String>? civilianMarkerOwnerIds,
+  }) {
+    return buildInitGameMapViewData(
+      game: game,
+      tileMapByRegion: tileMapByRegion,
+      topologyByRegion: topologyByRegion,
+      cellSize: cellSize,
+      seed: seed,
+      configSummary: configSummary,
+      greatPowerColorOverride: greatPowerColorOverride,
+      visibilityByTile: visibilityByTile,
+      warpLinks: warpLinks,
+      resourceExtractionUnitsByTile: resourceExtractionUnitsByTile,
+      resourceExtractionEffectiveUnitsByTile:
+          resourceExtractionEffectiveUnitsByTile,
+      resourceExtractionBlockedUnitsByTile: resourceExtractionBlockedUnitsByTile,
+      civilianMarkerOwnerIds: civilianMarkerOwnerIds,
+    );
+  }
+}
+
+/// Builds a [DualRegionViewScenario] from dual-region maps, topologies, and game.
+DualRegionViewScenario dualRegionViewScenario({
+  required Game game,
+  required TileMapResult oldWorldMap,
+  required TileMapResult newWorldMap,
+  required MapTopology oldWorldTopology,
+  required MapTopology newWorldTopology,
+}) {
+  return DualRegionViewScenario(
+    game: game,
+    tileMapByRegion: {'oldWorld': oldWorldMap, 'newWorld': newWorldMap},
+    topologyByRegion: {
+      'oldWorld': oldWorldTopology,
+      'newWorld': newWorldTopology,
+    },
+  );
+}
+
+/// Convenience wrapper for the common dual-region [buildInitGameMapViewData] call.
+InitGameMapViewData buildViewDataForScenario(
+  DualRegionViewScenario scenario, {
+  int cellSize = 8,
+  int? seed,
+  String? configSummary,
+  Map<String, (int r, int g, int b)>? greatPowerColorOverride,
+  Map<String, TileVisibility>? visibilityByTile,
+  List<WarpLink>? warpLinks,
+  Map<String, int>? resourceExtractionUnitsByTile,
+  Map<String, int>? resourceExtractionEffectiveUnitsByTile,
+  Map<String, int>? resourceExtractionBlockedUnitsByTile,
+  Set<String>? civilianMarkerOwnerIds,
+}) {
+  return scenario.buildViewData(
+    cellSize: cellSize,
+    seed: seed,
+    configSummary: configSummary,
+    greatPowerColorOverride: greatPowerColorOverride,
+    visibilityByTile: visibilityByTile,
+    warpLinks: warpLinks,
+    resourceExtractionUnitsByTile: resourceExtractionUnitsByTile,
+    resourceExtractionEffectiveUnitsByTile: resourceExtractionEffectiveUnitsByTile,
+    resourceExtractionBlockedUnitsByTile: resourceExtractionBlockedUnitsByTile,
+    civilianMarkerOwnerIds: civilianMarkerOwnerIds,
   );
 }
