+++ test/check_map_lib_file_size_test.dart
@@ -2,15 +2,16 @@ import 'dart:io';
 
 import 'package:test/test.dart';
 
-import '../tool/check_map_gen_file_size.dart';
+import '../tool/check_map_lib_file_size.dart';
 
 const _genRoot = 'packages/colonizethis_map/lib/src/gen';
 const _viewRoot = 'packages/colonizethis_map/lib/src/view';
+const _renderRoot = 'packages/colonizethis_map/lib/src/render';
 
 void main() {
-  test('passes for the real gen/view layers on dev', () {
+  test('passes for the real gen/view/render layers on dev', () {
     final logs = <String>[];
-    final code = runCheckMapGenFileSize(
+    final code = runCheckMapLibFileSize(
       Directory.current.path,
       info: logs.add,
       err: logs.add,
@@ -19,37 +20,38 @@ void main() {
       code,
       0,
       reason:
-          'Every colonizethis_map gen/view source file must stay at or below '
-          '${maxMapGenFileNonCommentLinesForTests()} non-comment lines.\n'
+          'Every colonizethis_map gen/view/render source file must stay at or below '
+          '${maxMapLibFileNonCommentLinesForTests()} non-comment lines.\n'
           '${logs.join('\n')}',
     );
   });
 
   test('the canonical scanned roots exist (gate cannot silently rot)', () {
-    for (final relativeRoot in mapGenFileSizeScanRoots) {
+    for (final relativeRoot in mapLibFileSizeScanRoots) {
       expect(
         Directory('${Directory.current.path}/$relativeRoot').existsSync(),
         isTrue,
         reason:
             'Scanned root $relativeRoot must exist; update '
-            'mapGenFileSizeScanRoots if the map gen/view layer moved.',
+            'mapLibFileSizeScanRoots if the map lib layer moved.',
       );
     }
   });
 
   test('fails when a gen file exceeds the 500 non-comment-line cap', () {
-    final temp = Directory.systemTemp.createTempSync('check_map_gen_size_fail_');
+    final temp = Directory.systemTemp.createTempSync('check_map_lib_size_fail_');
     addTearDown(() => temp.deleteSync(recursive: true));
 
     File('${temp.path}/$_genRoot/big_pass.dart')
       ..createSync(recursive: true)
       ..writeAsStringSync(List.filled(501, 'final x = 1;').join('\n'));
     Directory('${temp.path}/$_viewRoot').createSync(recursive: true);
+    Directory('${temp.path}/$_renderRoot').createSync(recursive: true);
 
     final logs = <String>[];
-    final code = runCheckMapGenFileSize(
+    final code = runCheckMapLibFileSize(
       temp.path,
-      scanRoots: const [_genRoot, _viewRoot],
+      scanRoots: const [_genRoot, _viewRoot, _renderRoot],
       info: logs.add,
       err: logs.add,
     );
@@ -60,23 +62,24 @@ void main() {
   });
 
   test('passes a gen file at exactly the cap (500 non-comment lines)', () {
-    final temp = Directory.systemTemp.createTempSync('check_map_gen_size_edge_');
+    final temp = Directory.systemTemp.createTempSync('check_map_lib_size_edge_');
     addTearDown(() => temp.deleteSync(recursive: true));
 
     File('${temp.path}/$_genRoot/edge_pass.dart')
       ..createSync(recursive: true)
       ..writeAsStringSync(List.filled(500, 'final x = 1;').join('\n'));
     Directory('${temp.path}/$_viewRoot').createSync(recursive: true);
+    Directory('${temp.path}/$_renderRoot').createSync(recursive: true);
 
-    final code = runCheckMapGenFileSize(
+    final code = runCheckMapLibFileSize(
       temp.path,
-      scanRoots: const [_genRoot, _viewRoot],
+      scanRoots: const [_genRoot, _viewRoot, _renderRoot],
     );
     expect(code, 0);
   });
 
   test('does not count comment-only lines toward the cap', () {
-    final temp = Directory.systemTemp.createTempSync('check_map_gen_size_cmt_');
+    final temp = Directory.systemTemp.createTempSync('check_map_lib_size_cmt_');
     addTearDown(() => temp.deleteSync(recursive: true));
 
     final lines = <String>[
@@ -87,22 +90,23 @@ void main() {
       ..createSync(recursive: true)
       ..writeAsStringSync(lines.join('\n'));
     Directory('${temp.path}/$_viewRoot').createSync(recursive: true);
+    Directory('${temp.path}/$_renderRoot').createSync(recursive: true);
 
-    final code = runCheckMapGenFileSize(
+    final code = runCheckMapLibFileSize(
       temp.path,
-      scanRoots: const [_genRoot, _viewRoot],
+      scanRoots: const [_genRoot, _viewRoot, _renderRoot],
     );
     expect(code, 0);
   });
 
   test('fails when a scanned root is missing (anti-rot existence check)', () {
-    final temp = Directory.systemTemp.createTempSync('check_map_gen_size_gone_');
+    final temp = Directory.systemTemp.createTempSync('check_map_lib_size_gone_');
     addTearDown(() => temp.deleteSync(recursive: true));
 
     final logs = <String>[];
-    final code = runCheckMapGenFileSize(
+    final code = runCheckMapLibFileSize(
       temp.path,
-      scanRoots: const [_genRoot, _viewRoot],
+      scanRoots: const [_genRoot, _viewRoot, _renderRoot],
       info: logs.add,
       err: logs.add,
     );
